requirements.txt
numpy
pandas
python-dotenv
plotly
matplotlib
openpyxl